+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './ui'
+export * from './link'
+export * from './theme/theme'
