+++ src/components/ui/header/index.ts
@@ -1,1 +1,1 @@
-export { default as Footer } from './Header'
+export { default as Header } from './Header'
