+++ src/components/ui/nft-card/index.ts
@@ -1,0 +1,1 @@
+export { default as NFTCard } from './NFTCard'
